新建用户接口以及详情接口
用户表与身份表联动

+++ demo-springboot-mysql/src/main/java/com/example/demo/user/UserRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends PagingAndSortingRepository<User, Integer> {
+public interface UserRepository extends PagingAndSortingRepository<User, Long> {
     Page<User> findByName(Pageable var1, @Param("name") String name);
 }
